+++ changelog.md
@@ -5,8 +5,9 @@
  - Support for tempFS
  - Improve support for 0.0.0.0 routes
  - Reworked the setting page for better clarity
+ - Added support for HTTP2 Cleartext (H2C)
  - Added advanced DNS options (challenge related, thanks @moham96)
- - Added trusted proxy option (Thanks @InterN0te)
+ - Added trusted proxy option for X-Forwarded-For (Thanks @InterN0te)
  - Reduce SmartShield false positive on the server panel by having two level of strictness (UI/panel vs. login)
  - Rclone rework, no more sub processes (increased performance and reliability)
  - Added the ability to manage Restic locks from the UI